services/reportes_service/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-dotenv
pandas
reportlab
openpyxl
